+++ src/shared/sleep-util.c
@@ -36,7 +36,7 @@
 #include "macro.h"
 #include "path-util.h"
 #include "siphash24.h"
-#include "sleep-config.h"
+#include "sleep-util.h"
 #include "stat-util.h"
 #include "stdio-util.h"
 #include "string-table.h"
